requirements.txt
streamlit==1.29.0
openai>=1.6.1,<2.0.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
chromadb==0.4.22
sentence-transformers==2.7.0
requests==2.31.0
kaggle==1.5.16 